% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blueconic.R
\name{blueconic}
\alias{blueconic}
\title{A blueconic Function}
\usage{
blueconic(df)
}
\arguments{
\item{data}{formatted as Blueconic data from Hackthon 2019}
}
\description{
This function allows you to predict product type
}
\examples{
blueconic()

}
\keyword{blueconic}
