% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{blueconic_model_demo}
\alias{blueconic_model_demo}
\title{blueconic model_demo}
\format{a \code{\link{xgb}} object}
\source{
Allianz 2019
}
\usage{
blueconic_model_demo
}
\description{
Simple example model that predicts product type using XGB
}

